moved serialization to inout
